COMMIT-3

+++ src/main/java/com/training/magang/ecommerce/product/service/ProductService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface ProductService {
 
     Product create(Product product);
-    Product findById(int id);
+    Product findById(Long id);
     List<Product> findAll();
     Product update(Product product);
-    Product delete(int id);
+    Product delete(Long id);
 
 }
